Optimisers

CurrentModule = NNJulia.Optimisers

NNJulia.Optimisers.AbstractOptimiser — Type

AbstractOptimiser

Every optimiser struct is a subtype of AbstractOptimiser

NNJulia.Optimisers.GradientDescent — Type

GradientDescent(lr::Float64)
GradientDescent()

Represents the vanilla gradient descent optimiser. The default constructor initialise the learning rate at 0.1

Field

lr: The learning rate

NNJulia.Optimisers.update! — Function

update!(opt::GradientDescent, model::AbstractModel)

Update the parameters of the model using the given optimiser.