Utils

NNJulia.split_train_test — Function

split_train_test(xData::AbstractArray, yData::AbstractArray, at::Float64)

Split the xData and yData (shuffled) into a train and a test set.

For example, with at=0.8, the train set will contain 80% of the original data and the test set will contain 20% of it.

NNJulia.plot_decision_boundary — Function

plot_decision_boundary(model::AbstractModel, xData::AbstractArray, yData::AbstractArray, levels::Int64=3)

Plot a decision frontier of the model for 2D data with the given number of levels.

NNJulia.to_one_hot — Function

to_one_hot(x::AbstractArray)

Convert an array to to one hot encoded format.